+++ src/modules/dateTimePicker/components/index.ts
@@ -2,3 +2,4 @@ export * from './dateTimePeriodPickerBase';
 export * from './dateTimePicker/dateTimePicker.component';
 export * from './dateTimePicker/dateTimePicker.interface';
 export * from './dayPicker/dayPicker.component';
+export * from './monthPicker/monthPicker.component';